+++ src/soap/xsd/XSD_QName.cpp
@@ -21,7 +21,7 @@ AXIS_CPP_NAMESPACE_START
         MinLength* minLength= getMinLength();
         if (minLength->isSet())
         {
-            if (strlen(value) < minLength->getMinLength())
+            if (strlen(value) < (unsigned int) minLength->getMinLength())
             {
                 AxisString exceptionMessage =
                 "Length of value to be serialized is shorter than MinLength specified for this type.  Minlength = ";
@@ -43,7 +43,7 @@ AXIS_CPP_NAMESPACE_START
         MaxLength* maxLength = getMaxLength();
         if (maxLength->isSet())
         {
-            if (strlen(value) > maxLength->getMaxLength())
+            if (strlen(value) > (unsigned int) maxLength->getMaxLength())
             {
                 AxisString exceptionMessage =
                 "Length of value to be serialized is longer than MaxLength specified for this type.  Maxlength = ";
